app.c: fix memory free on static data
When the two binding information structs are initialized, default
values are given. Since a pointer pointing to .data is also valid,
g_free() cannot distinguish whether it can be free'd or not.

The memory corruption bug is triggered when no configuration file
is given, no configuration file on the default location, and a
command line option is given.

This commit removes default values as initialization (replaced with
NULL, so g_free() can ignore them), and set them later in the
command line parser.